oAuth -- add comments, package renaming, minor cleanup (#1225)
oAuth -- add comments, package renaming, minor cleanup
