+++ scripts/test_suite.args
@@ -23,7 +23,7 @@
 --add-exports org.igv/org.broad.igv.feature.genome.fasta=junit
 --add-exports org.igv/org.broad.igv.feature.tribble=junit
 --add-exports org.igv/org.broad.igv.feature.tribble.reader=junit
---add-exports org.igv/org.broad.igv.google=junit
+--add-exports org.igv/org.broad.igv.oauth=junit
 --add-exports org.igv/org.broad.igv.gwas=junit
 --add-exports org.igv/org.broad.igv.lists=junit
 --add-exports org.igv/org.broad.igv.maf=junit
