+++ src/main/java/org/broad/igv/prefs/PreferencesEditor.java
@@ -3,8 +3,7 @@ package org.broad.igv.prefs;
 import org.broad.igv.logging.*;
 import org.broad.igv.DirectoryManager;
 import org.broad.igv.Globals;
-import org.broad.igv.google.OAuthUtils;
-import org.broad.igv.sam.AlignmentTrack;
+import org.broad.igv.oauth.OAuthUtils;
 import org.broad.igv.ui.IGV;
 import org.broad.igv.ui.color.ColorSwatch;
 import org.broad.igv.ui.color.ColorUtilities;
