+++ src/main/java/org/broad/igv/ui/action/LoadFromURLMenuAction.java
@@ -32,9 +32,9 @@ package org.broad.igv.ui.action;
 import org.broad.igv.logging.*;
 import org.broad.igv.exceptions.HttpResponseException;
 import org.broad.igv.feature.genome.GenomeManager;
-import org.broad.igv.google.GoogleUtils;
-import org.broad.igv.google.OAuthProvider;
-import org.broad.igv.google.OAuthUtils;
+import org.broad.igv.util.GoogleUtils;
+import org.broad.igv.oauth.OAuthProvider;
+import org.broad.igv.oauth.OAuthUtils;
 import org.broad.igv.prefs.Constants;
 import org.broad.igv.prefs.PreferencesManager;
 import org.broad.igv.session.SessionReader;
