+++ src/main/java/org/broad/igv/util/AmazonUtils.java
@@ -3,8 +3,8 @@ package org.broad.igv.util;
 import com.google.gson.JsonObject;
 import org.broad.igv.Globals;
 import org.broad.igv.aws.IGVS3Object;
-import org.broad.igv.google.OAuthProvider;
-import org.broad.igv.google.OAuthUtils;
+import org.broad.igv.oauth.OAuthProvider;
+import org.broad.igv.oauth.OAuthUtils;
 import org.broad.igv.logging.LogManager;
 import org.broad.igv.logging.Logger;
 import org.broad.igv.ui.IGVMenuBar;
